smarter log cleanup script (#533)

+++ .openshift/cron/daily/log-cleanup.sh
@@ -1,1 +1,7 @@
-rm -rf $OPENSHIFT_JBOSSAS_LOG_DIR/*.log.*
+if [ $OPENSHIFT_JBOSSAS_LOG_DIR ]; then
+	rm -rf $OPENSHIFT_JBOSSAS_LOG_DIR/*.log.*
+fi
+
+if [ $OPENSHIFT_JBOSSEAP_LOG_DIR ]; then
+	rm -rf $OPENSHIFT_JBOSSEAP_LOG_DIR/*.log.*
+fi
